Be nice in the presence of method aliasing
Flushing each fetched AST from the cache makes it impossible to re-fetch
the ast for a method under a different name. This change disables this
automatic flushing. The effect on memory should be minimal, as even in
the presence of flushing the ASTs still stick around, bound to their
methods.